requirements.txt
asgiref==3.2.10
boto3==1.15.5
botocore==1.18.5
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
dj-database-url==0.5.0
Django==3.1
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==6.0.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.10.1
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
Pillow==7.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
